requirements.txt
fastapi 
uvicorn 
pandas 
numpy 
psycopg2-binary 
python-dotenv 
pathlib
typing
sqlalchemy
dotenv